Align icons in tables marginally better.
The negative margins on images shift icons down one pixel, presumably
because of rounding or something.  They're not necessary in tables, so
we can at least ditch them there.

Maybe images should default to margin: 0, and then only images in dls
(plus whatever other edge cases where having an image stretch the line
throws it out of alignment with something else) can have the negative
margins?

To some extent though vertically-centered icons are just doomed to look
too low when the Pokémon are so low in the frame.  Hmmmm.