Shrink the banner to fit the window when necessary.
This helps slightly on mobile, at least in the default BlackBerry
browser.  Before, it would zoom out by default so that the banner would
fit.

The site's styling is simple enough that I'd prefer to keep one
stylesheet that works on desktop and mobile.  It's not there, but I
don't think it will take much.