Deny permissions in the right place.
This is left over from before Hacker School, so I don't know exactly
what the bug was, but I'm guessing it had to do with permissions on a
Pokémon being denied when they shouldn't have?  Or something??  Whyyy
did I not commit this when I did it?

In any case it looks right and I'm pretty sure I remember it fixing the
bug so I'm gonna commit it.