Keep track of Pokémon's birthdays and origins.
This is shit but it works for now.  I had to get the birthday column up
today if I wanted to keep track of birthdays from the beginning.

I'll have to set a lot of was_from_hacks manually.  Oh well.

Also add the previous alembic script because I'm the WORST