Prevent <th>es from wrapping mid-word, and widen the Speed header.
For some people it was saying:

Spee
d

Apparently slapping word-wrap: break-word on everything wasn't the best
solution; this isn't the only place it's happening, but it's the only
prominent one that I know of.  (The other is the Buy Pokémon/Items cart
table, and only if you have a balance of at least four digits.)  I'll
have to figure out some better CSS later; I just want to quickly fix
this for now.