Let admins ban users.
Banned users will have a notice on their pages stating why they were
banned.  They will be filtered out of the trainer list, and their
Pokémon will be filtered out of the Pokémon list and species pages'
census lists.

They will be automatically logged out, and if they try to log in again,
they will be notified that they were banned, and why.  This seemed a
lot simpler than having to remember to account for banned members in
permissions everywhere.

Unbanning is not possible for the time being.  I'll implement it if and
when we need it.