Give Pokémon sprites a circle backdrop instead of a frame.
A lot of them looked off-center in the frame.  For example, Mawile's
body is centered, but its horns loll off to one side, so the sprite as
a whole is off-center.  A bunch of Pokémon do this sort of thing.

The frame makes them look unambiguously off-center, whereas the circle
puts the focus on whatever part actually *is* in the center, and makes
the composition look more deliberate.

I also wanted to ditch the ridge border because different browsers
render it slightly differently.  Firefox makes the light edges lighter
and the dark edges darker, so I picked a saturated blue that would look
good after getting watered down.  However, other browsers actually use
the assigned colour, which is too intense.