Properly sanitize Markdown output.
Okay!  Remember in the commit message for 22b646c where I was all
fascinated that Mako was escaping just the raw HTML, and leaving
Markdown HTML alone?  It turns out that's not really what was
happening.  It was escaping the input *before* passing it on to
Markdown, and since that left most Markdown syntax alone, it *looked*
like it was doing the right thing.

However, {move:King's Shield} broke because it escaped the apostrophe,
and if anyone wanted to use > to produce a blockquote, that wouldn't
have worked either.  Also, like I said, that didn't prevent JavaScript
links.

So now I'm using an actual HTML sanitization library, bleach, which
solves all of the above problems.  I may have forgotten to whitelist
some element or another, though.  I guess we'll see.