Show moves' actual base power instead of rounded base damage.
The last digit is actually relevant (and has been for months) now that
the Damage & Energy Guide says not to round until the end.  I wanted to
streamline the whole process and switch over to base damage, but it's
not worth pushing it right now I guess.  If it ever does happen it'll
be part of a big overhaul and I'll switch back then.