Let admins close any battle and edit closing info.
I want to let any mod set the battle back to a "needs closure" state and
go back to the closing form, which would already be filled out with what
the ref entered so that the mod could just tweak it.

It's also confusing that you can enter a KO count without checking the
Pokémon off and I need to fix that.  Maybe the checkbox is unnecessary?
I think when I first put this together I was worried that people would
enter 0 for Pokémon that didn't battle, and I figured making a separate
checkbox would be more explicit.  But I'm pretty sure the actual upshot
of this is that people are forgetting to check the box and that's why we
periodically see battles where supposedly no Pokémon were used.

But anyway I am super scrambled this afternoon so for now I'll just make
it possible for me to fill out the closing form again even when the
battle is already awaiting approval.