Pretty up the trainer edit form.
It now uses fieldsets, and it's all one form, aside from the password
reset and ban forms (which are now under their own h1, labelled the
"danger zone").

I'm thinking that having fieldsets which are obvious boxes, and then
having the save button outside of the fieldsets, makes it more intuitive
that the save button applies to all the fieldsets.  (Also it's pretty!)

Of course, nobody else will get to appreciate this, but there are
probably other forms that could benefit from fieldsets.  And maybe soon
I'll get trainer editing working smoothly enough to let mods use it too?
(Mostly I was just inspired to use fieldsets all of a sudden, and this
was the best form to play around with them on.)

Side note: I should go through and hammer down the palette at some
point.  In particular, the blue I borrowed for this looks a little weird
against the slightly-gray background.