Finish the trade schema.
I did most of this back in, like, March.  I don't remember why I bumped
the SQLA version, but I'm going to assume it was necessary.

I think I'm not going to require mod approval for trades.  Instead, I'm
going to require you to get in a battle before you can trade, and put a
warning up saying that if you abuse the trade system you'll get banned.
I think that will be enough.  In any case, it'll mean I can get this up
for Christmas faster.

Holy shit the entire text of this commit message is justified, I didn't
even do that on purpose I swear.  Not even this paragraph.