Let trainers deposit their entire squad. #41
Trading made it possible to empty your squad, and trying to withdraw
after that caused a 500 — the validator on the *deposit* form was
borking because it ended up having a maximum of -1.  (The form wasn't
even being validated; it just borked when trying to *construct* it.)

I never really had a good reason for requiring people to keep a Pokémon
in their squad at all times, though, so I figured the easiest way to
deal with that bug was just to drop that requirement.