Update Pokémon form order.
I did this separately to keep the last commit's diff readable.  There's
a gap before Magearna so that it won't have to move when I add the
spoilery Pokémon.

I'd like it to sort Alola Forms as if they were separate families, and I
guess for consistency it should do the same for e.g. Shellos and
Deerling, but for now I'm just going to keep doing it the old way.