Update propagate_moves.sql for Alola Forms.
This is pretty gross lmao.  It associates forms with the corresponding
prevo form based on form_order, defaulting to the prevo's form 1.  This
is technically wrong in a handful of places, but none that actually make
a difference — e.g. it means that Mothim only inherits moves from Plant
Burmy, but whatever.