Constrain Pokémon portraits to 120×120.
The ones from the Sun/Moon Pokédex are 256×256.  A lot of them look kind
of silly scaled down to 120×120, and a lot of them just plain don't fit
with the OR/AS ones, but I'll figure that out later.

(n.b. the sprites aren't actually stored in the repo.)