Add non-rippable categories to new moves.
- Direct healing:   Shore Up, Floral Healing, Purify
- Indirect healing: Pollen Puff
- Exact damage:     Nature's Madness
- Biting:           Psychic Fangs
- Ballistics:       Beak Blast

I'm not sure which healing category to put Pollen Puff in; it heals
allies based on the damage it *would* do.  I think indirect makes more
sense, but direct might work better for balance purposes.  I've asked
for opinions in the social group, but I'm putting it in indirect for
now.